Refactor client components
