+++ app/_lib/components/TourItem.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 
 export default function TourItem(
