--- app/tourism/TourList.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-export default function TourList(
-  {
-    items,
-  }: Readonly<{
-    items: React.JSX.Element[],
-  }>): React.JSX.Element {
-  return (
-    <div className="tour-list">
-      {items}
-    </div>
-  );
-}
